pkg/util: move httpstream to k8s.io/apimachinery
