+++ cmd/kube-aggregator/pkg/apiserver/BUILD
@@ -61,7 +61,6 @@ go_library(
         "//pkg/genericapiserver/registry/generic/rest:go_default_library",
         "//pkg/genericapiserver/registry/rest:go_default_library",
         "//pkg/genericapiserver/server:go_default_library",
-        "//pkg/util/httpstream/spdy:go_default_library",
         "//pkg/util/workqueue:go_default_library",
         "//pkg/version:go_default_library",
         "//vendor:github.com/golang/glog",
@@ -69,6 +68,7 @@ go_library(
         "//vendor:k8s.io/apimachinery/pkg/apis/meta/v1",
         "//vendor:k8s.io/apimachinery/pkg/labels",
         "//vendor:k8s.io/apimachinery/pkg/runtime",
+        "//vendor:k8s.io/apimachinery/pkg/util/httpstream/spdy",
         "//vendor:k8s.io/apimachinery/pkg/util/runtime",
         "//vendor:k8s.io/apimachinery/pkg/util/wait",
         "//vendor:k8s.io/apiserver/pkg/endpoints/request",
