+++ cmd/kube-aggregator/pkg/apiserver/handler_proxy.go
@@ -22,12 +22,12 @@ import (
 	"sync"
 
 	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/util/httpstream/spdy"
 	genericapirequest "k8s.io/apiserver/pkg/endpoints/request"
 	restclient "k8s.io/client-go/rest"
 	"k8s.io/client-go/transport"
 	"k8s.io/kubernetes/pkg/genericapiserver/endpoints/handlers/responsewriters"
 	genericrest "k8s.io/kubernetes/pkg/genericapiserver/registry/generic/rest"
-	"k8s.io/kubernetes/pkg/util/httpstream/spdy"
 
 	apiregistrationapi "k8s.io/kubernetes/cmd/kube-aggregator/pkg/apis/apiregistration"
 )
