+++ pkg/client/unversioned/portforward/BUILD
@@ -18,7 +18,7 @@ go_library(
     deps = [
         "//pkg/api:go_default_library",
         "//pkg/kubelet/server/portforward:go_default_library",
-        "//pkg/util/httpstream:go_default_library",
+        "//vendor:k8s.io/apimachinery/pkg/util/httpstream",
         "//vendor:k8s.io/apimachinery/pkg/util/runtime",
     ],
 )
@@ -31,8 +31,8 @@ go_test(
     deps = [
         "//pkg/client/unversioned/remotecommand:go_default_library",
         "//pkg/kubelet/server/portforward:go_default_library",
-        "//pkg/util/httpstream:go_default_library",
         "//vendor:k8s.io/apimachinery/pkg/types",
+        "//vendor:k8s.io/apimachinery/pkg/util/httpstream",
         "//vendor:k8s.io/client-go/rest",
     ],
 )
