+++ pkg/genericapiserver/registry/generic/rest/proxy.go
@@ -30,12 +30,12 @@ import (
 	"time"
 
 	"k8s.io/apimachinery/pkg/api/errors"
+	"k8s.io/apimachinery/pkg/util/httpstream"
 	utilnet "k8s.io/apimachinery/pkg/util/net"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	genericfeatures "k8s.io/apiserver/pkg/features"
 	utilfeature "k8s.io/apiserver/pkg/util/feature"
 	"k8s.io/apiserver/pkg/util/proxy"
-	"k8s.io/kubernetes/pkg/util/httpstream"
 
 	"github.com/golang/glog"
 	"github.com/mxk/go-flowrate/flowrate"
