+++ pkg/util/BUILD
@@ -62,7 +62,6 @@ filegroup(
         "//pkg/util/framer:all-srcs",
         "//pkg/util/goroutinemap:all-srcs",
         "//pkg/util/hash:all-srcs",
-        "//pkg/util/httpstream:all-srcs",
         "//pkg/util/i18n:all-srcs",
         "//pkg/util/initsystem:all-srcs",
         "//pkg/util/interrupt:all-srcs",
