+++ staging/src/k8s.io/apimachinery/pkg/util/httpstream/spdy/upgrade.go
@@ -22,7 +22,7 @@ import (
 	"strings"
 
 	"k8s.io/apimachinery/pkg/util/runtime"
-	"k8s.io/kubernetes/pkg/util/httpstream"
+	"k8s.io/apimachinery/pkg/util/httpstream"
 )
 
 const HeaderSpdy31 = "SPDY/3.1"
